Refactor runner into subpackage
