+++ src/nf_meta/__main__.py
@@ -1,7 +1,7 @@
 import click
 from functools import wraps
 from nf_meta.engine.errors import GraphValidationError, ValidationError, format_errors_for_cli
-from nf_meta.engine.runner import run_metapipeline, Runners, NfMetaRunnerError
+from nf_meta.runner import run_metapipeline, Runners, NfMetaRunnerError
 from nf_meta.engine.graph import MetaworkflowGraph
 from nf_meta.engine.session import start_session
 from nf_meta.editor import start_editor_backend
