+++ src/nf_meta/runner/cascade_runner.py
@@ -1,0 +1,5 @@
+
+
+class NfCascadeRunner:
+    def __init__(self):
+        raise NotImplementedError("Runner not implemented")
