+++ src/nf_meta/runner/errors.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class NfMetaRunnerError(Exception):
+    message: str
+   
+    def __str__(self) -> str:
+        return self.message
