+++ src/nf_meta/runner/python_runner.py
@@ -1,3 +1,8 @@
+from .errors import NfMetaRunnerError
+
+from nf_meta.engine.graph import MetaworkflowGraph
+from nf_meta.engine.models import Workflow, GlobalOptions
+
 import hashlib
 import logging
 import os
@@ -8,12 +13,8 @@ import uuid
 import threading
 from collections import deque
 from contextlib import contextmanager
-from dataclasses import dataclass
-from enum import StrEnum
-from typing import Optional, Protocol
+from typing import Optional
 from pathlib import Path
-from nf_meta.engine.graph import MetaworkflowGraph
-from nf_meta.engine.models import Workflow, GlobalOptions
 
 import yaml
 from rich.live import Live
@@ -24,66 +25,6 @@ from rich.console import Console, Group
 
 logger = logging.getLogger(__name__)
 console = Console()
-
-class Runners(StrEnum):
-    PYTHON = "python"
-    PLATFORM = "platform"
-    DAISY_CHAIN = "daisy_chain"
-    NEXTFLOW_MONO = "nextflow_mono"
-
-
-class Runner(Protocol):
-    def run(self, graph) -> None: ...
-    def resume(self, graph) -> None: ...
-
-
-def run_metapipeline(
-        g: MetaworkflowGraph,
-        runner_name: Runners = Runners.PYTHON,
-        resume = False,
-        verbose = True,
-        output_lines = 20,
-        start: str = None,
-        target: str = None,
-        profile: str = None
-    ) -> None:
-    logger.info("Started runner")
-
-    errors = []
-    if start:
-        if not g.get_workflow_by_id(start):
-            errors.append(f"Starting workflow {start} is not a valid workflow id!")
-        else:
-            logger.info(f"Starting workflow: {start}")
-        
-    if target:
-        if not g.get_workflow_by_id(target):
-            errors.append(f"Target workflow {target} is not a valid workflow id!")
-        else:
-            logger.info(f"Target workflow: {target}")
-    
-    if errors:
-        raise NfMetaRunnerError("\n".join(errors))
-
-    runner = None
-    match runner_name:
-        case Runners.PYTHON:
-            runner = SimplePythonRunner(output_window_size=output_lines, start=start, target=target, extra_profile=profile)
-        case _:
-            raise NotImplementedError("Requested runner not implemented yet")
-    
-    if resume:
-        runner.resume(g)
-    else:
-        runner.run(g)
-
-
-@dataclass
-class NfMetaRunnerError(Exception):
-    message: str
-   
-    def __str__(self) -> str:
-        return self.message
 
 
 class SimplePythonRunner:
